chore: remove old js modules
