+++ babel.config.cjs
@@ -1,6 +1,7 @@
 module.exports = {
   presets: [
     ["@babel/preset-env", { targets: { node: "current" } }],
-    ["@babel/preset-react", { runtime: "automatic" }]
+    ["@babel/preset-react", { runtime: "automatic" }],
+    "@babel/preset-typescript"
   ]
 };
