--- src/NodeEditorContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react'
-
-const NodeEditorContext = createContext({ updateNodeText: () => {} })
-
-export default NodeEditorContext
